Don't remove patch lines from spec in testbuild
Without this patch, the testbuild CI job may fail the obs validators if
the package has patches to apply[1]. This causes the testbuild job to
fail and hence the mkcloud job to never start. This patch keeps the
patch directives in the spec file so they can be applied and the
validators can pass.

[1] https://ci.suse.de/job/cloud-crowbar-testbuild-pr/4500/console