openstack-mkcloud: reduce default nodenumber for ha cloud to 3
Previously the default and minimum nodenumber for an hacloud was 4.
The default cluster only used two nodes though. Having one compute
node by default should be sufficient.

In order to free or save resources on our workers we now lower the
number of nodes for an hacloud to 3 (2 for the cluster, one compute).

All current hacloud jobs define their required nodenumber explicitly:
  grep -lr hacloud=1 | xargs grep -E "nodenumber=|hacloud="
So the existing jobs should not be affected by this change.