drop IFS manipulation, keep IFS unset
In case the mtu is different than exptected the for loop
manipulated the IFS (to match full lines). After the loop
the IFS was tried to set to its original value.
But an unset IFS behaves differently than an empty IFS.

So lots of commands that were run after this manipulation
failed (most prominent "ssh $sshopts ..." which resulted in
errors like: "command-line line 0: garbage at end of line").

This PR fixes it by dropping the need for the IFS manipulation.