qa_crowbarsetup: mark PTF key workaround for =< cloud7
The fix to import the PTF key is in crowbar.
But it is missing in Cloud6 GM and Cloud7 GM.
So we need to keep this workaround here until
Cloud7 is EOL.

Therefore the workaround is now in an if clause,
that marks it to only apply to =< Cloud7. So now its
obvious, that we can refactor it out when Cloud7 is EOL.
Sideeffect: the revert PR (waiting until C7 EOL) can be closed.