qa_crowbarsetup: fix deletion of testvm
This is a fix for f87bca5fe912c2c5ce1eb7180d585674bb6bad81
The $instanceid variable was not set so the wait_for loop
did not wait for anything.

This PR fixes it as it queries the instanceid of a potentially
running instance called testvm first and only calls delete on
it if an instance exists.