devstack: added option to unstack
Since we do not use swift-replicas, devstack runs swift in foreground
to save up ram, by not starting all process, as a result jenkins does
not consider the job as completed since the ssh is never closed as
terminal is still held causing jenkins to stall forever.

By explicitly doing an unstack we get a chance to terminate openstack
processes in an orderly manner.

we do not want the job to fail if something fails in cleanup step so
added a '|| :', this makes sure we have exit 0 even if clean up fails.