Make crm failure checks optional
In non-CI mkcloud deployments, it might be acceptable to continue with
the proposal stage even if pacemaker has detected service failures. This
patch adds a want_crm_failcount_skip variable which can be set to 1 to
turn off the default fail-fast behavior in HA clouds. This is only
applicable in the proposal step. In the testsetup step we specifically
disallow skipping this check.