Recover fencing during reboot
We found in some cases the reboot step take too long for pacemaker, and
fencing is triggered. We cannot really control the time it takes to
reboot the node so this proposal takes care of recovery from fencing.

The steps followed are for the particular case of fencing during reboot.
I have not considered other cases.