Fix ipmi node reboot for ipmi_sim
The ipmitool selftest command does not work properly when the ipmi
interfaces are being simulated by ipmi_sim, resulting in the error:

 Bad response: (Invalid command)

though the interfaces are otherwise responsive and functional. Since
this use of selftest is being used for checking the IPMI password,
replace it with the info command since that also requires a password and
can be used to check that a password works.