Stop using direct dist.suse.de mounts for test updates
We meanwhile mirror the -testupdates to clouddata, so at least
they're consistent and stable and don't change at random point
in times during deployment, breaking everything. Also it avoids
a NFS mount to dist, which is unavailable for anyone outside
the NUE office network, which makes remote work signficantly harder.