Add support for simulating BMCs with ipmi_sim
ipmi_sim implements the IPMI 1.5 and 2.0 LAN protocols, so we can use it
in combination with the BMC support in QEMU to simulate IPMI controllers
in libvirt nodes.

This patch creates veth interfaces attached to the cloud bridge to act as BMC
devices, then runs ipmi_sim in screen sessions that bind to one end of
the veth pairs, allowing us to route traffic on the standard BMC port
(623) to different IP addresses on the BMC network defined in
network.json.

Using this feature requires the OpenIPMI package to be installed on the
host system.

Co-authored-by: Bernhard M. Wiedemann <bwiedemann@suse.de>