Refactor add_mount
In order to prepare `add_mount` to be able to digest clouddata hirearchy
through `localreposdir_src` we need to modify it's parameters - detach
the hardcoded nfs server part so the path part is available inside
`add_mount`.  With this change `add_mount` will be able to use that path
value and set up a local mount on the same path.