Rename cachedir to cache_dir
This is more consistent with the other variables like log_dir. Also
move it out of /tmp, as it is really bad behavior to install into
a global location that is not safe for modification from concurrent
runs, as that will inevitably cause terrible, hard to trace
corruption bugs otherwise.