Add cache_clouddata=1 support
This is a quick way to enable consistent caching of the clouddata
data on the host system and enforce local-only usage of those
when installing a virtual cloud. This is very useful for unstable
and slow VPN connections (e.g. road warriors) as it doesn't require
downloading gigabytes of data every time or causing machines to
hang due to remote NFS mounts that are timing out.

just do export cache_clouddata=1 and be happy.