Disable magnum in develcloud8
The magnum barclamp currently fails to deploy
in develcloud8 because some packages are missing
from Devel:Cloud:8 (python-kubernetes and
python-marathon).

This commit temporarily disables magnum in develcloud8
in order to get Jenkins jobs to pass over the point
where they try to deploy magnum and fail.

This is required to unblock the mkcloud8 gating job.