CI: Document the log parser rules
This patch adds a short note in the jenkins section of the README about
the log parser rules file. Having this here would have helped me when I
was grepping for 'Parsed Console Output' which is what the jenkins view
is titled. This does not link back to the
"openstack-mkcloud-log-parser-test" page of our "keep jenkins green"
document since that is in a private repository and this repository is
open source.