qa_crowbarsetup: Hide stderr from nova commands
When the nova client fails something, it produces an ugly ERROR message
that the jenkins log parser treats as a potential error. We don't need
to treat 'show' commands that are checking the existence of something
for idempotency purposes as failures. This also restructures some delete
commands to first check for existence before attempting the delete,
because if the delete fails for some reason other than nonexistence we
want to know about it.

When the openstack client command fails something, the stderr doesn't
contain ERROR and isn't treated as such by the log parser, so we might
as well let it keep being verbose.