Don't sync with delete on Provo site
The provo site is a bit broken atm, we don't have a proper ibs
mirror in provo, so some stuff is coming manually from nue. In
order for the two rsyncs to not interfere, the sync script doesn't
sync with delete and there is an extra script on the host to sync
the remainders with a --delete.