fix database engine selection
If people are deploying cloud8+ we want mysql to be the engine. We
already added a rule for that but we did not export the variable. This
meant that the `want_database_sql_engine` variable was not displayed by
env and not forwarded to onadmin calls, so for qa_crowbarsetup.sh the
variable was still undefined - leaving us with an mkcloud that wants to
deploy drdb on a three-node cluster in some cases.

```
> env | grep variable
> variable=3
> env | grep variable
> export variable=3
> env | grep variable
variable=3
```