ironic: Add deployment support
This patch adds the necessary network.json changes for an ironic
deployment as well as an 'ironic' case in the proposal deployment step
and updates to the nova proposal step to support ironic. Deployment of
the ironic barclamp is guarded by the $want_ironic setting which is
already in use elsewhere.

For now this leaves the existing ironic jenkins jobs alone (which uses
the batch command to customize the network and ironic barclamps) but we
can consider switching to this model, which is more in line with how
other barclamps are deployed, in the future.