ironic: Stop using pxe_ssh driver (SCRD-768)
The pxe_ssh driver was deprecated in Ocata and removed in Pike[1]. This
patch removes it from the cloud8 job.

We aren't currently testing baremetal deployments with the pxe_ssh
driver so removing it is not a loss. In order to fully test ironic it
would be better to get virtualized deployments to work with the
pxe_ipmitool driver.

[1] https://docs.openstack.org/releasenotes/ironic/pike.html#id10