Optionally allow vendor change for cleanvm
Depending on the provenance of the VM image used by cleanvm, it may have
packages beyond the JeOS set of packages installed already. If one of these
packages exists in Devel:Cloud:* in a newer change, a vendor change may be
needed. In this case zypper will prompt for that vendor change, even in
non-interactive mode, thus stalling deployment.

This commit introduces the $allow_vendor_change variable. If set to '1' it,
allows vendor changes cleanvm to prevent the problem outlined above from
happening. This needs to be set explicitly. By default vendor changes are
_not_ allowed.