Skip setting of CLOUDTESTISO for GM7/GM7+up
These packages are not installable for GM7+up anymore as the
testiso remains on GM level. furthermore the actual usage of
those packages is commented out so there is little value in
trying to make them installable in the first place. To be
revisited when the magic Cloud 8 testing plan is in place.