Add want_ses_version
Allow a user to override the SES version installed. Currently in mkcloud
the SES version is determined by the version of SLES you are running:

  https://github.com/matthewoliver/automation/blob/external_ceph/scripts/lib/mkcloud-common.sh#L548-L564

This patch makes is possible for a user to override what version of
SES will be installed, well really it says what SES repos to include.
This will be useful for testing SES versions, but very important for
the SAP squad as it allows us to use SLES 12-SP2 combined with SES5.

To use it,  just set it to the SES version:

  export want_ses_version=5