Adding the deployexternalceph step
The deployexternalceph step will deploy SES using deepsea on a bunch of
lonelynodes. This means, in order to run this step you must first run
setuplonelynodes and crowbar_register.

This patch introduces a new environment variable:

  - want_external_ceph

want_external_ceph, tells mkcloud to not deploy internal
ceph but still setup the nodes with the SES repositories etc.
So to use this, you need something like:

  export nodenumberlonelynode=4
  export want_external_ceph=1
  mkcloud setuplonelynodes crowbar_register deployexternalceph

Optionally if you want to override the SES version:

  export nodenumberlonelynode=4
  export want_external_ceph=1
  export want_ses_version=5
  mkcloud setuplonelynodes crowbar_register deployexternalceph

This will deploy SES5 on 4 lonely nodes that will have been registered
with crowbar.

By default, the SES networks for the ceph cluster to mkcloud cloud will be:

  public = public
  cluster (ceph replication) = storage

If however you specify $want_separate_ceph_network:

  export want_separate_ceph_network=1

Then the networks will be:

  public = ceph
  cluster (ceph replication) = storage

This step:
  - Adds the storage, public and optionally the ceph network to the
    nodes
  - Makes sure the required SES repo is enabled on each node
  - Sets their aliases to be ceph0 - cephN
  - Deploys SES on the cluster nodes

Noting that the SES uses deepsea (salt) to deploy the cluster, as such
node ceph0 will become the salt master.

The top of the build_ses.sh script contains global variables for the
number of different ceph roles to be deployed on the nodes:

    NUM_MASTER=1
    NUM_ADMIN=1
    NUM_MON=3
    NUM_MDS=1
    NUM_IGW=0
    NUM_RGW=2

Most of these roles will be randomly assigned between the nodes, with some
exceptions, such as at least 1 of the master and admin roles will always be
on ceph0. This is so you can easily find the controlling node.

Note want_ceph and want_external_ceph are currently mutually exclusive.