mkcloud: allow to run supportconfig as onhost step
In order to collect supportconfig data from a successful run
we need the possibility to collect supportconfig data.
Doing this for all builds is not ideal (too much data).
So we want to trigger a weekly run that captures supportconfig
data. This PR adds this support via the onhost+stepname mechanism.