Add more information to wait_tftpd
The onadmin_wait_tftpd function waits for the tftpd server to be ready
on the admin node. The wait_for call does not include a process name
which leads to the following output on failure:

      Ran into error: Waiting for 'unknown process' timed out.

Since the above message is not very informative by itself, this change
adds some more information to the wait call.