mkcloud-common: fix is_onadmin and is_oncontroller functions
The function is_onhost checks if mkcloud is in the list of
${BASH_SOURCE[@]} which includes all sourced files.

The functions is_onadmin and is_oncontroller only checked
$BASH_SOURCE which only holds one file.

To be able to check in a function in mkcloud-common if it is
running onadmin is_onadmin needs to check the entire list of
sourced files. This is because $BASH_SOURCE will only state
that the code came from mkcloud-common.sh - which does not say
where the code is executed.