Add obscpio support to trackupstream
Basically the script needs to adjust to look for the new
service name ("obs_scm" instead of "tar_scm") and treat files
ending in *obscpio similar to existing tar file matching logic.

Last point is that the recompress service might not exist, so
make that one optional and try to guess differently if it
is not configured.