Install Pool and Updates repos on develcloud*
The motivation for this patch is that currently develcloud6 builds fail
when SSL is turned on because the SLES12-SP1-Updates repo does not
contain the fix for [1], where the Cloud-6-Updates package
does[2]. It is reasonable to install the Pool and Updates repos on
development builds since we want to mimic how users' systems will be
set up. We can rely on package versioning to ensure that if packages
exist in both Updates and Devel repositories, the ones from Devel will
be installed.

[1] https://bugzilla.suse.com/show_bug.cgi?id=1043607
[2] https://trello.com/c/AOI554Pf