Move the ceph onadmin tests to ceph nodes
This patch fixes bug 1053012, where the qamcloud's ceph tests
fail due to a routing problem from the crowbar node to the
public interface on the ceph nodes.

If deploying ceph the testsetup step had some ceph and radowgw
tests that would run on the admin node. These tests have now
been moved to ceph nodes.

The ceph tests are run on the first monitor node and the
radosgw tests are run on the first radosgw node, where if
network routing is a problem.. then there really is a
deployment problem.