github_pr: config for cloud automation
This is the github_pr[1] configuration for
for the repos
  SUSE-Cloud/automation
  SUSE-Cloud/cct

This tool together with this configuration will replace the
github-status[2] tool and parts of the trigger jobs[3] that launch
the pull request CI builds.

The configuration is needed to filter the pull requests that
anybody can create for the public automation repo.
The filters are extracted from github-status[2] and transformed
to the configuration - most as yaml, but also also as ruby code, which
is included via the configuration file.

In short:
- github_pr[1] is the new generic tool that can handle any github repo
- this commit holds the filter logic of github-status[2] for [3]

[1] https://github.com/openSUSE/github-pr
[2] https://github.com/SUSE-Cloud/automation/blob/master/scripts/github-status/github-status.rb
[3] https://github.com/SUSE-Cloud/automation/blob/master/jenkins/ci.suse.de/cloud-automation-pr-trigger.yaml
    https://github.com/SUSE-Cloud/automation/blob/master/jenkins/ci.suse.de/cloud-cct-pr-trigger.yaml