qa_crowbarsetup: Remove ostestr usage
This commit removes ostestr command usage from qa_crowbarsetup.sh.
Because ostestr doesn't have 'last' subcommand[0] but ostestr just
ignore the 'last' subcommand, actually. However, it just runs same tests
again. It just wastes the resource. So I think switching to use testr is
better.

In the upstream OpenStack development, we're switching to use stestr
instead of testr. So, it's better to switch to use it in the future.

[0] https://docs.openstack.org/os-testr/latest/user/ostestr.html