ardana: Make clouddata-server configurable
provo-clouddata seems to be outdated currently (eg ardana-ansible is
from 2017-12-14 but the version on clouddata is from 2018-01-09). To
be able to use the most recent pkg versions, make the clouddata server
configurable.
Note: The switch to the provo-clouddata server happened in
c35e3761b057cd7 . That's why the discover of outdated packages
happened now.