ardana: Move lifecycle-manager endpoint the Management network
With commit c5858c0390c we switched the packager to serve repos in the
Management network. This change is reflected that also in the
network-groups attribute to have ardana create the IP Tables rules for
the packager/lifecycle-manager port (79 by default) for the correct IP
Address.

That also me that until we find a way to declare static IP allocations
in the input mode for all networks we're hosting everything in the
Management network.

Co-Authored-By: Thomas Bechtold <tbechtold@suse.com>