mkcloud: use e2fsck instead fsck
This is to prevent fsck from incorrectly getting a different file system
type from fstab. This would happen when a file system in fstab is
mounted by label, with the same label in use by the fs to fsck but a
different type.

A different way around this would be to set the env FSTAB_FILE to the
empty string for fsck invocation. This more generic solution is not
necessary as resize2fs is used on the same fs, which like e2fsck only
works on the ext fs types.

Using fsck -t ext4 will not work when the type is declared in fstab as
that has higher priority.

Example log of the problem this fixes:
mkcloud# ./mkcloud.sh plain
...
resize2fs /dev/mapper/cloud-cloud.admin1 failed! (safelyret=1) Aborting.
...
mkcloud# grep xfs /etc/fstab
/dev/disk/by-label/ROOT / xfs defaults 1 1
mkcloud# fsck -V -y -f /dev/mapper/cloud-cloud.admin1
fsck from util-linux 2.29.2
[/sbin/fsck.xfs (1) -- /] fsck.xfs -y -f /dev/mapper/cloud-cloud.admin1
/sbin/fsck.xfs: XFS file system.