SCRD-2756 post-deployment sanity checks
Check that Ardana doesn't mess with the filesystem inappropriately by
removing or modifying rpm-owned files, or adding files into /usr.
This is done by running rpm -Va and rpm -qf before and after Ardana
deployment (including the tempest run, if deployment succeeded), and
comparing the results.  If any tampering is discovered, init.yml will
fail with helpful errors.  As a last step it will also report which
(if any) of site.yml, tempest, and these checks failed, so that
multiple failures don't get hidden behind one of the failures.

Whitelists are used to ignore known exceptions, some of which need
to be fixed in the future.

A playbook is also provided for rapid repeated testing of the
post-deployment checks, without having to deploy an entire fresh cloud
each time.