pr trigger: switch crowbar PR trigger to github_pr
After testing github_pr as for the automation and cct
repo we now switch the trigger for the crowbar repos
to github_pr as well as combining them to one job
(crowbar org and sap-oc org).

The merge into one job will save api requests as
data that is fetched for both orgs is now cached.