syncSLErepos: go back to syncing isos and local unpacking
There seems to be a race with the OBS repopublisher, which first
syncs the rpms and then repodata, which causes the 2nd rsync
to freak out. Go back to syncing the iso and then loopback
sncing it. for that to be of reasonable speed, the bwlimit
needs to be removed.