testbuild: log rebuild links to easier rerun a job
From time to time triggered jenkins jobs are not processed.
This might be due to various reasons.

In order to easier rebuild a missing job the testbuild
script will now log a rebuild link. The link will open
a confirmation page in jenkins and prefill all parameters.
The job is not starting automatically, one additional
button klick is needed (and wanted as protection).

The one who starts the rebuild must make sure there are no
identical parallel jobs running. If there are identical jobs
the status of the job reporting last will be the status of
that commit on github.