github_pr: refactor logging of JenkinsJobTrigger action
Add the ability to log details about a just triggered job
to the JenkinsJobTriggerAction itself.

There is a separate logging Action available (which is also
used here) but it makes more sense to log directly instead
of moving the logging to a separate Action in the filter
chain.

Due to this, the logging can be enhanced with information
that is only relevant for the JenkinsJobTrigger action.
These enhancements will be added with the next commit.