update_automation: allow to override automation repo
For several QA jenkins jobs the feature was used to override
the location of the automation repo. This was used to stage
code changes before they became a pull request.

This feature was accidentally dropped with the latest
refactoring of update_automation.

This commit adds back this feature but with a slightly different
usage.
The definition must follow this pattern:
     <domain>/<owner>/<repo>#<branch>"
e.g. github.com/SUSE-Cloud/automation#master"

The definition of the branch can be omitted (incl. the "#").
Then the branch name will use "master" as default.