ardana: Increase inactivity timeout even further
When init.yml is running site.yml there is not log activity for a very
long time, which cause the jenkins job to timeout for the larger models.

In the mid term we need to find a better solution here. This should be
considererd a temporary fix to address the immediate issue of all jobs
timeing out.