mkcloud: enable mkcloud options specific for CI
We may want mkcloud to act differently on CI runs than on manual
executions. A good example for this would be the timing reports at the
end of the mkcloud run which make more sense only in a CI environment.

We introduce the variable CI_RUN: when enabled, mkcloud should consider
this a CI run. The first use of this variable will be for only enabling
timing reports on manual runs. Note that timing is still profiled, but
not reported.