jjb: fix submit task for ardana packages in submit job
In the last refactoring an error sneaked in with commit
bdf9bf63a9648090ac8a56f775e4808a633d05c3

The unwanted behaviour is that ardana packages should not
be submitted to non-Staging project.

The wanted new behaviour (in PR #2429) after the refactoring
was that ardana packages are also submitted to OBS.

These changes fix these issue.