mkcloud: Wrapped SKIPSUPPORTCONFIG in function
The function setuphost was called with SKIPSUPPORTCONFIG, but since
setuphost is part of the regex in line 1532 (in ## MAIN ##) error_exit
is not called on setuphost errors, thus the config wasn't collected in
the first place.

The exceptions to this are the two calls to pvcreate and vgcreate
inside the function libvirt_do_setuphost(), which were decorated with
safely, which calls error_exit. But since SKIPSUPPORTCONFIG was set, it
wouldn't collect the configs either.

I removed the SKIPSUPPORTCONFIG on the setuphost call and added a
safely_skip_support decorator function, which provides the same functionality,
but without the config collection, which should make the code more
readable.