Fix routing to external network for std-min case
The ardana-ci model expects to create a bond0 for -mgmt out of eth1
and eth3, and eth4 is supposed to be the external network. Due to
a typo external-mgmt were swapped, which caused awkward networking
faults. rename _unused to _mgmt2 to make that bond clearer, and
disable vlan for external network as the neutron router needs to
be able to route it.