ardana-job: optionally reboot compute nodes before post-deploy checks
Add a new reboot_computes option to reboot all compute nodes before
running tempest and the rpm checks.  This will add a few minutes to
the run, so it's disabled by default.

Previously we were implicitly relying on set -e behaviour for aborting
immediately when site.yml fails.  This is somewhat obfuscated and
brittle, so also be more explicit about it.