mkcloud: default to x86_64 only when not using NUE mirror
The non-NUE clouddata mirrors are hosting x86_64 only, so
it makes sense to default to that.

Apparently users of the provo VPN endpoint are stumbling over
not setting to be x86_64 only (even if they only want to use x86_64).
So add some magic to detect the case and stop mounting non-x86_64
mointpoints in that case.