ardana: Allow to drop in arbitrary input model changes
This moves our adjustments of the input model into a separate
subdirectory and renames the files to their standard names as used in
the ardana-ci models.

The init.yml playbook is adjust to copy all files in that subdirectory
onto the destination system instead of naming the files explicitly. This
should make it easier to add modifications to the models.