ardana: Add heat templates to deploy SES in an OpenStack cloud
This might be useful when testing the external SES integration for
Ardana in the engcloud (but can be used in any cloud running Heat).

Use the stack with:
openstack stack create --parameter "ses_key_name=tbechtold" \
    --parameter "ses_extra_network=a-neutron-net" \
    -t heat-ses-cluster.yaml  --wait ses-stack-name

When calling heat-ses-cluster.yaml, the "ses_key_name" is
mandatory and the used ses_image_id (defaults to jeos-SLE12SP3) needs
to contain cloud-init.
After the stack creation, you can login via the floating ip (see
"openstack stack show ses-stack-name") and watch the deployment
process of the ceph cluster via:
tail -f /var/log/cloud-init-output.log

The "ses_extra_network" is a Neutron network ID (or name if
unique). Every node in the Ceph cluster will be connected to this
network so this is the integration point between an Ardana deployment
and a SES deployment.
The IP addresses of the MON and OSD nodes are available as Heat stack
output parameters.