force upgrade existing python packages
There seems to be some existing packages in travis which may be outdated
or missing requirements. In this case requests seems to be installed
but not urllib3.

A simple pip install wont update those existing packages leaving the
environment kind of broken for our tests. So instead mark the pip
install with -U so it upgrades the existing  packages, which in turn
installs the dependencies correctly.