Correct & improve mkcloud doc
I've added a few things to this document based on my recent
experience in getting it set up for the first time. I've also
taken the liberty of modifying the script so that it will be
closer to working out of the box for most people. They will
still need to edit paths to taste.