github_pr: allow pr checks from non-fork branches
Usually developers fork a repository and create PRs from their
forks to the upstream repository.
For SUSE/cloud-specs this is a bit different.

To allow PR checking for all PRs even if they came from the a
branch of the repository itself (instead of a fork) we hereby
allow PRs from the user SUSE (which is the organization user).