openstack-ardana: fix build pool cleanup
Fixes the cleanup logic used to determine the list
of old stacks in a stack pool which need to be
deleted when a job begins.

The current logic fails to cover cases where the number
of old jobs is lower than the maximum number set for the
job pool. The effect is that a new job deletes all old
stacks that it finds.

Replacing tail with awk helps better encode the logic of
selecting "all but the first X-1, chronologically sorted
stacks" used to determine which stacks need to be deleted.