openstack-ardana: fix subnet heat resource deps
When a ResourceGroup is used to group multiple servers
together, the implicit dependency between those servers
and the list of networks they are attached to no longer
works [1].

The result of this is we get intermittent errors when
deleting heat stacks, because subnets are deleted before
the servers that are using them.

The workaround (and recommendation [2]) is to use a subnet
list or a port list instead of a network list when defining
OS::Nova::Server resources.

[1] https://bugs.launchpad.net/heat/+bug/1547435
[2] https://bugs.launchpad.net/heat/+bug/1243992