openstack-ardana: heat template generator
Implements an ansible playbook, role and required ansible modules that take in an
existing Ardana input model definition coupled with an optional virtual setup
configuration descriptor and generates a complete heat orchestration template
that can be used to deploy the supplied input model in a virtual environment.

The included README.md file includes usage instructions, as well as detailed
information on the implementation itself.