Add Gerrit Change ID param to openstack-ardana job
This change adds the ability to point the openstack-ardana Jenkins job
at an open change proposal on gerrit.suse.provo.cloud, build it on IBS,
and use the resulting repository as an update repository for Ardana
integration tests.

It adds a new job, cloud-ardana-testbuild-gerrit, which uses the
openstack-trackupstream node label since it requires IBS credentials.
This job uses a python script, build_test_package.py, which borrows
heavily from the rpm-packaging project's createproject.py script to
create a test project in IBS, copy a package into it, point the source
service at a locally updated version of the source, and build the
package. The repository has the publish flag set so it can be used by
the integration test as an additional repository on top of the regular
Cloud repository that is mirrored in Provo. By default, it creates a
subproject of the opensuseapibmw's home project to test in.

It adds a conditional build step to the openstack-ardana job that
triggers the new job if the gerrit_change_id parameter is set. For now,
users can manually set this to test a change they want feedback on, but
in the future this could be automatically set by a trigger that reacts
to new changes in Gerrit.