Allow specifying multiple gerrit change IDs
To check out multiple changes, add each ID in a comma-separated list.
The project that gets created to test them will have a name that
includes each ID, separated by a dash. If multiple changes are for the
same repository, the order in which they are given in the parameter is
the order in which they will be merged.