allocpool: match cmd name when pid found
In case allocpool finds a process with the pid of the pid file,
also the cmdline needs to be matched to verify, if this process
was an mkcloud run.
Otherwise it might happen (and it did) that the pid counter runs
over and restarts and a pidfile now holds a pid of a different
process (in our case a deamon). This would block a slot from
allocpool even though there no longer is an mkcloud process.

This change does the comparision of the cmdline via a variable
in the environment. By this allocpool remains independent and
does not need to know about mkcloud itself.