openstack-ardana: fix network subnet ranges
The subnet ranges configured in the input model for
the ARDANA-NET and MANAGEMENT-NET networks didn't match
those used in the heat template. This caused the newer
version of configuration processor to complain with the
following type of message:

 "ERR: Server ardana-ccp-c0-m1 (192.168.245.4) using
 interface model ARDANA-INTERFACES does not have a
 connection to a network which contains its address."

, which basically means that the addresses assigned
to servers in servers.yml (i.e. allocated by heat/openstack)
don't match the address ranges configured in the net_global.yml
file.