openstack-ardana: reduce ansible verbosity
Using -vvv is flooding the logs with ansible connection
information which has never proven useful in the past.
Reducing verbosity to make the output easier to decipher
(-v should be sufficient for most purposes).

Also removing the heat template generator debug task,
because that info is already covered by the -v modifier.