Make GM6/GM7 cloudsource more reliable
The fetching of ISOs was hardcoded to always sync from
download.suse.de/install. However the SUSE IT team is frequently
cleaning up that directories, which means ISO's that we rely on
are disappearing and breaking our CI. Any attempt at getting them
restored is time consuming. So instead we'll now host those isos
that we rely on on clouddata. Some refactoring and renaming of
parameters makes what is happening a bit more obvious.