Handle edge cases of cross-repo dependencies
Support building packages with cross-repo dependencies using the format
Depends-On: $change_url.

A full URL to the gerrit change is required. No support for change ID's
in the format of I\w{32} is available. Instead the unambiguous change
number from the URL should be used.

This commit also adds checks for already merged changes, wrong branch
dependencies, and dependencies in the middle of a commit tree.