Adjust scenarios for change of default database backend
The default database backend for Cloud 7 will soon change from
"postgresql" to "mariadb". In order to keep the scenarios deployable
with GA as well as with GA+Updates after the changed default, we're
forcing the backend to "postgresql" in the crowbar batch files for now.

Many of the batch files are still taylored for postgresql anyway (e.g.
using two-node cluster with and without drbd)