Switch Cloud9 HA jobs to use ssl
The HA+SSL combination isn't very well tested in our CI, but it
is the more common scenario that customers / PoCs are using. For
example the SAP Cloud Project is requiring this combination, but
also other users are choosing this combination. Hence we should
be testing it.

For now only switch Cloud9. Switching more once we have more
testing passes.