Move trackupstream scripts from job to files
The trackupstream shell scripts were inlined in the job configs. Move
them as is into stand alone files. Except indentation whitespace
corrected and the part that updates the automation repo omitted.

Create a symlink in the ci.suse.de part to the macros.yaml from
opensuse, to be able to share these. The builder update-automation from
the opensuse part is used to replace the duplicate code in the suse
part.