Run openstack-mkcloud job on changes to its script
In #2709 we moved the bulk of the openstack-mkcloud job to a script
within the automation repo, but that new script was not added to
the file filter for the openstack-mkcloud job. Add the new script
directory to the file filter so that the openstack-mkcloud job runs
on changes to its builder script.

Move the pull-request checkout to its own script to be run before the
main builder script so that the builder script can be self-gated.

Depends on work from #2713 and #2687