SCRD-5000 ignore depends on repos that don't have RPMS
There are  repos that don't have RPMs, these should be
ignored in the RPM based tests. Right now we see the issue with a-d-t, but
in the furtue it is possible that we have product changes that
require infr changes to land. It would be good to be able to record
that code dependency. We may also see this bewteen test/QA repos
and changes in product code.