Add Jenkins API python script: jenkins-job-cancel
Adds a new `jenkins-job-cancel` Python script which can be
used to abort all but the most recent build belongint to a job
and matching a list of parameter values.

This script can be used by jobs triggered for SCM events,
such as GitHub and Gerrit, to abort builds running for obsolete
change versions (e.g. GitHub branch contents or Gerrit patches
that are no longer current etc).