keystone_password: Increase crowbar batch timeout
In an HA deployment a chef-client run can take around 10 minutes on a
single node. As the crowbar needs to wait for existing chef-client runs
to finish before the new chef-client run for applying the keystone
barclamp can be triggered we need to increase the timeout of the crowbar
batch call to have enough time for 2 chef-client runs + some safety
margin. This commit increases the timeout from 15 to 25 minutes.