openstack-ardana: readd heat stack create/delete job
Moves the logic dealing with long operations involving creating/deleting
heat stacks in the ECP cloud to a separate Jenkins job, openstack-ardana-heat.
The advantages of having this handled in a separate job:
 - ECP infrastructure related failures can be monitored separately
 - the job can be run manually to cleanup stacks left over from previous
 openstack-ardana pipeline runs
 - cleaning up a heat stack during the post-build actions of other jobs
 (e.g. openstack-ardana) can be triggered and run without blocking the
 upstream job, which means that upstream jobs can finish earlier and
 report the result faster
 - last but not least, the job can be reused by other pipelines which
 need to manage heat stacks, for example a job that automates creating a
 SES cluster

The heat stack management tasks protected by the ECP-API resource
lock have also been reduced to those that actually create or delete
heat stacks, while the monitoring tasks are executed outside this
lock. This effectively reduces the time that jobs spend waiting
for the ECP-API resource to become available.