openstack-ardana: run ardana-ssh-keyscan.yml playbook for cloud9 on hardware
Cloud 9 sets `host_key_checking = True` on ansible.cfg which requires
accepting the host keys when is the firt time running the playbook on a
new host. The keys are also automatically added to known hosts when
running site.yml which calls ardana-ssh-keyscan.yml playbook.

For hardware deployments we need to run wipe_disks.yml before site.yml
and as wipe_disks.yml does not call ardana-ssh-keyscan.yml it get stuck
waiting for input to accept the ssh keys.

This change adds ardana-ssh-keyscan.yml to the list of playbooks that
are ran for cloud9 hardware deployment so we make sure the hosts keys are
on known hosts file before running wipe_disks.yml