openstack-ardana: fix heat template file name
When running openstack-ardana job the default is to use the input model
from ardana repository (model parameter is set and scenario_name is
empty).

For the heat template file name the same logic is used, by default use
the model value and only use scenario_name when its defined and not
empty.